+++ src/components/App/filters/filters.js
@@ -1,11 +1,6 @@
-import { unCamelCase } from '../../../utils/un-camel-case'
 import moment from 'moment'
 import { KYC_REQUEST_STATES } from '@/constants'
-
-export function filterAccountType (accountType) {
-  const newStr = unCamelCase(accountType)
-  return newStr.replace(/Account Type /g, '')
-}
+import config from '@/config'
 
 export function filterDateWithTime (date, format = 'DD MMM YYYY [at] hh:mm:ss') {
   try {
@@ -21,4 +16,16 @@ export function localizeIssuanceRequestState (type) {
 
 export function cropAddress (value) {
   return `${value.slice(0, 4)}…${value.slice(-4)}`
+}
+
+export function roleIdToString (roleId, postProcessor) {
+  return {
+    [config.ACCOUNT_ROLES.notVerified]: 'Unverified',
+    [config.ACCOUNT_ROLES.general]: 'Individual',
+    [config.ACCOUNT_ROLES.corporate]: 'Corporate'
+  }[roleId]
+}
+
+export function lowerCase (string) {
+  return String(string).toLowerCase()
 }